server/requirements.txt
python-dotenv
polars==0.16.3
torch==2.0.0+rocm6.2
transformers==4.31.0
datasets
sentencepiece==0.1.99
accelerate==0.21.0
huggingface_hub==0.17.3
evaluate==0.4.0
pandas
nltk
absl-py
rouge_score